+++ GaugeSupport.php
@@ -22,6 +22,14 @@
 
 class GaugeSupportPlugin extends MantisPlugin {
 
+	const MANTISGRAPH = 'MantisGraph';
+	const MANTISGRAPH_VERSION = '2.25.0';
+
+	/**
+	 * @var bool $mantisgraph_loaded True if MantisGraph plugin is available
+	 */
+	private $mantisgraph_loaded = false;
+	
 	function register() {
 		$this->name = plugin_lang_get( 'title' );
 		$this->description = plugin_lang_get( 'description' );
@@ -30,6 +38,11 @@ class GaugeSupportPlugin extends MantisPlugin {
 		$this->requires = array(
 			'MantisCore' => '2.0.0',
 			);
+
+		# MantisGraph required so we don't need to bundle chart.js ourselves
+		$this->uses = array(
+			self::MANTISGRAPH => self::MANTISGRAPH_VERSION,
+		);
 
 		$this->author = "Cas (based upon Charly Kiendl's work), Damien Regad";
 		$this->contact = 'Cas@nuy.info';
@@ -47,10 +60,63 @@ class GaugeSupportPlugin extends MantisPlugin {
 	
 	function hooks() {
 		return array(
+			'EVENT_LAYOUT_RESOURCES' => 'resources',
 			'EVENT_MENU_MAIN' => 'menuLinks',
 			'EVENT_MENU_ISSUE' => 'issueVoteLink',
 			'EVENT_VIEW_BUG_EXTRA' => 'renderBugSnippet',
 		);
+	}
+
+	/**
+	 * Initialize plugin
+	 *
+	 * Check soft dependency on MantisGraph plugin.
+	 */
+	public function init() {
+		if( plugin_is_registered( self::MANTISGRAPH ) ) {
+			$t_version_check = plugin_dependency(
+				self::MANTISGRAPH,
+				self::MANTISGRAPH_VERSION
+			);
+			$this->mantisgraph_loaded = $t_version_check == 1;
+		}
+		if( !$this->mantisgraph_loaded ) {
+			log_event( LOG_PLUGIN, $this->missingMantisGraph() );
+		}
+	}
+
+	/**
+	 * @return string
+	 */
+	public function missingMantisGraph() {
+		return sprintf(
+			plugin_lang_get( 'mantisgraph_missing' ),
+			self::MANTISGRAPH_VERSION
+		);
+	}
+
+	/**
+	 * Include javascript for chart.js if needed.
+	 *
+	 * Scripts are only loaded on pages that need them (currently, view.php).
+	 *
+	 * @return void
+	 */
+	function resources() {
+		$t_page = basename( $_SERVER['SCRIPT_FILENAME'] );
+		if( $this->mantisgraph_loaded && $t_page == 'view.php' ) {
+			$t_mantisgraph = plugin_get( self::MANTISGRAPH );
+			$t_mantisgraph->include_chartjs();
+
+			echo "\t", '<script src="' . plugin_file( "GaugeSupport.js" ) . '"></script>', "\n";
+		}
+	}
+
+	/**
+	 * @return bool True if ChartJs library is available;
+	 */
+	public function isChartJsAvailable() {
+		return $this->mantisgraph_loaded;
 	}
 
 	function menuLinks($p_event) {